+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/FlexibleSearchElementType.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,6 +18,5 @@
 package com.intellij.idea.plugin.hybris.flexibleSearch
 
 import com.intellij.psi.tree.IElementType
-import org.jetbrains.annotations.NonNls
 
-class FlexibleSearchElementType(debugName: @NonNls String) : IElementType(debugName, FlexibleSearchLanguage)
+class FlexibleSearchElementType(debugName: String) : IElementType(debugName, FlexibleSearchLanguage)