+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/FlexibleSearchTokenType.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,10 +18,9 @@
 package com.intellij.idea.plugin.hybris.flexibleSearch
 
 import com.intellij.psi.tree.IElementType
-import org.jetbrains.annotations.NonNls
 import java.util.regex.Pattern
 
-class FlexibleSearchTokenType(debugName: @NonNls String) : IElementType(debugName, FlexibleSearchLanguage) {
+class FlexibleSearchTokenType(debugName: String) : IElementType(debugName, FlexibleSearchLanguage) {
 
     override fun toString() = super.toString()
         .takeIf { it.isNotBlank() }