+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/highlighting/FlexibleSearchHighlighterColors.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,7 +20,6 @@ package com.intellij.idea.plugin.hybris.flexibleSearch.highlighting
 
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
 import com.intellij.openapi.editor.colors.TextAttributesKey
-import org.jetbrains.annotations.NonNls
 
 object FlexibleSearchHighlighterColors {
 
@@ -47,7 +47,7 @@ object FlexibleSearchHighlighterColors {
     val FXS_OPERATION_SIGN = key("FXS_OPERATION_SIGN", DefaultLanguageHighlighterColors.OPERATION_SIGN)
 
     private fun key(
-        externalName: @NonNls String,
+        externalName: String,
         fallbackAttributeKey: TextAttributesKey
     ) = TextAttributesKey.createTextAttributesKey(externalName, fallbackAttributeKey)
 }