+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/highlighting/PolyglotQueryHighlighterColors.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,7 +20,6 @@ package com.intellij.idea.plugin.hybris.polyglotQuery.highlighting
 
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
 import com.intellij.openapi.editor.colors.TextAttributesKey
-import org.jetbrains.annotations.NonNls
 
 object PolyglotQueryHighlighterColors {
 
@@ -37,7 +37,7 @@ object PolyglotQueryHighlighterColors {
     val PGQ_TYPE = key("PGQ_TYPE", DefaultLanguageHighlighterColors.FUNCTION_DECLARATION)
 
     private fun key(
-        externalName: @NonNls String,
+        externalName: String,
         fallbackAttributeKey: TextAttributesKey
     ) = TextAttributesKey.createTextAttributesKey(externalName, fallbackAttributeKey)
 }