+++ src/com/intellij/idea/plugin/hybris/toolwindow/system/bean/tree/BSTree.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -29,7 +29,6 @@ import com.intellij.ui.TreeUIHelper
 import com.intellij.ui.tree.AsyncTreeModel
 import com.intellij.ui.treeStructure.Tree
 import com.intellij.util.containers.Convertor
-import org.jetbrains.annotations.NonNls
 import java.io.Serial
 import javax.swing.event.TreeModelListener
 import javax.swing.tree.DefaultMutableTreeNode
@@ -56,7 +55,7 @@ class BSTree(val myProject: Project) : Tree(), DataProvider, Disposable {
         }, SEARCH_CAN_EXPAND)
     }
 
-    override fun getData(dataId: @NonNls String) = null
+    override fun getData(dataId: String) = null
 
     override fun dispose() = Unit