+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### `Project Import` enhancements
 - Loading backoffice sources provided by SAP on project import [#1096](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1096)
 
+### Other
+- Migrated to `IntelliJ Platform Gradle Plugin 2.x` [#1099](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1099)
+
 ## [2024.1.1]
 
 ### `Project Import` enhancements